+++ CHANGES.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.0] - 2020-03-04
+### Added
+- Support for using pylint installed in a virtual environment.
+See `force_venv_activation` parameter on `inithook` method.
 
 ## [2.0.0] - 2019-10-19
 ### Added